Split TooltipElement
